requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
black==25.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
defusedxml==0.7.1
deprecation==2.1.0
filelock==3.18.0
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.0
mypy_extensions==1.1.0
packaging==25.0
parsel==1.10.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
postgrest==1.1.1
Protego==0.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.3.0
pytz==2025.2
queuelib==1.8.0
realtime==2.5.3
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tldextract==5.3.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
w3lib==2.3.1
websockets==15.0.1
zope.interface==7.2
